Stellingen table gemaakt v3

+++ files/js/home.js
@@ -13,6 +13,8 @@ if (document.getElementById('button-partijen')) {
     })
 }
 
+
+
 function showPage(page) {
     if (currentPage !== page) {
         let request = new XMLHttpRequest();
@@ -22,11 +24,15 @@ function showPage(page) {
 
                 if (page === "stellingen") {
                     document.getElementById("button-partijen").classList.remove('active');
+                    document.getElementById("button-"+page).classList.add('active');
                 } else if (page === "partijen") {
                     document.getElementById("button-stellingen").classList.remove('active');
+                    document.getElementById("button-"+page).classList.add('active');
+                } else {
+                    document.getElementById("button-stellingen").classList.remove('active');
+                    document.getElementById("button-partijen").classList.remove('active');
                 }
 
-                document.getElementById("button-"+page).classList.add('active');
                 currentPage = page;
 
                 assignButtonFunctions();
@@ -38,16 +44,34 @@ function showPage(page) {
 }
 
 function assignButtonFunctions() {
-    document.querySelectorAll('.editTest').forEach((element) => {
-        element.addEventListener('click', () => {
-            editEntry();
-        })
-    });
-    document.querySelectorAll('.deleteTest').forEach((element) => {
-        element.addEventListener('click', () => {
-            deleteEntry();
-        })
-    });
+
+    if (document.getElementsByClassName("editTest")[0]) {
+        document.querySelectorAll('.editTest').forEach((element) => {
+            element.addEventListener('click', () => {
+                editEntry();
+            })
+        });
+    }
+
+    if (document.getElementsByClassName("deleteTest")[0]) {
+        document.querySelectorAll('.deleteTest').forEach((element) => {
+            element.addEventListener('click', () => {
+                deleteEntry();
+            })
+        });
+    }
+
+    if (document.getElementById('button-add-stelling')) {
+        document.querySelector('#button-add-stelling').addEventListener('click', () => {
+            showPage("stelling-add");
+        });
+    }
+
+    if (document.getElementById('button-add-partij')) {
+        document.querySelector('#button-add-partij').addEventListener('click', () => {
+            showPage("partij    -add");
+        });
+    }
 }
 
 function editEntry() {
